# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Between the Lines bookstore API",
    "version": "1.0.0",
    "description": "Simulated commerce only. Same-origin HttpOnly sessions. Money is integer USD cents. Send application/json for mutations. Hosted in one instance."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/config": {
      "get": {
        "summary": "Public demo configuration",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "storeName": {
                      "type": "string"
                    },
                    "mode": {
                      "type": "string"
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "simulated": {
                      "type": "boolean"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "liveSearch": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "storeName",
                    "mode",
                    "categories",
                    "simulated",
                    "currency",
                    "liveSearch"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/v1/sessions": {
      "post": {
        "summary": "Create or resume session; reset deletes the old session",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "messages": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Message"
                      }
                    },
                    "cart": {
                      "$ref": "#/components/schemas/Cart"
                    },
                    "mode": {
                      "type": "string"
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Order"
                      }
                    }
                  },
                  "required": [
                    "messages",
                    "cart",
                    "mode",
                    "orders"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reset": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": []
              }
            }
          }
        }
      }
    },
    "/api/v1/books": {
      "get": {
        "summary": "Browse curated shelves or search saved records without an upstream request",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "books": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Book"
                      }
                    },
                    "warning": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "books"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxPriceCents",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "inStockOnly",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 24,
              "default": 12
            }
          }
        ]
      }
    },
    "/api/v1/discover": {
      "get": {
        "summary": "Search saved records and live Open Library; validated editions receive fictional demo offers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "books": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Book"
                      }
                    },
                    "page": {
                      "type": "integer"
                    },
                    "nextPage": {
                      "type": "integer"
                    },
                    "warning": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "books",
                    "page"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200
            }
          },
          {
            "in": "query",
            "name": "kind",
            "schema": {
              "enum": [
                "all",
                "title",
                "author",
                "topic"
              ],
              "default": "all"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 20,
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "maxPriceCents",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/v1/books/{id}": {
      "get": {
        "summary": "Read a seeded edition",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^OL[0-9]+M$"
            }
          }
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "summary": "List only the current session’s orders",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "400": {
            "description": "HTTP 400: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "HTTP 401: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "HTTP 403: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "HTTP 404: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "HTTP 409: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "HTTP 429: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "HTTP 503: validation, session, conflict, rate limit or provider error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bookshop_session"
      }
    },
    "schemas": {
...
      "Cart": {
        "type": "object",
        "properties": {
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CartLine"
            }
          },
          "totalCents": {
            "type": "integer"
          },
          "count": {
            "type": "integer"
          },
          "currency": {
            "const": "USD"
          },
          "version": {
            "type": "integer"
          }
        },
        "required": [
          "lines",
          "totalCents",
          "count",
          "currency",
          "version"
        ]
      },
...
      "Message": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "user",
              "assistant"
            ]
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object"
                }
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  }
}